Change the wrongAnswer CSS.
Add new state in in dctionary - isWrong, that store the data, if you make a mistake.

Change the following function:
fetchData : add setIsWring

Add new Funciton:
setWrong: Manages state changes in case of a wrong response

Implements wrong, and setIsWrong in the subcomponent
Add new useState method, that it flashes the wrong answer for a few seconds, then returns to its default state.